requirements.txt
pandas
selenium
python-dotenv
pytesseract
pillow
opencv-python